Inserindo o update no MarketController

+++ backend/app/Controllers/Http/MarketController.js
@@ -1,9 +1,9 @@
 'use strict'
 
-const Market = use('App/Models/Market')
 const Address = use('App/Models/Address')
 const Category = use('App/Models/Category')
 const MarketRepository = use('App/Repository/MarketRepository')
+const AddressRepository = use('App/Repository/AddressRepository')
 
 /** @typedef {import('@adonisjs/framework/src/Request')} Request */
 /** @typedef {import('@adonisjs/framework/src/Response')} Response */
@@ -30,41 +30,27 @@ class MarketController {
   async store ({ request, response }) {
     const market = await this.marketRepository.create(request.body)
 
-    const address = new Address()
-    address.number = request.body.address.number
-    address.complement = request.body.address.complement
-    address.neighborhood = request.body.address.neighborhood
-    address.city = request.body.address.city
-    address.uf = request.body.address.uf
-    address.country = request.body.address.country
-    address.cep = request.body.address.cep
-    address.street = request.body.address.street
+    if( !market ) {
+      return response.status(500).json({message: "Market wasn't created"})
+    }
+
+    const addressRepository = new AddressRepository()
+    const address = await addressRepository.create(request.body.address)
+
+    if( !address ) {
+      await market.delete()
+      return response.status(500).json({
+        message: "Address wasn't created. Market deleted."
+      })
+    }
 
     const categories = {
       categories: request.body.categories,
       registred: [],
     }
 
-    try{
-      await market.save()
-    }
-    catch (err) {
-      return response.status(500).json({
-        'message': "Market wasn't created"
-      })
-    }
-
-    try{
-      await address.save()
-      await market.address().associate(address)
-      market.address = address
-    }
-    catch (err) {
-      await market.delete()
-      return response.status(500).json({
-        'message': "Address error! Market deleted"
-      })
-    }
+    await market.address().associate(address)
+    market.address = address
 
     //Funcionalidade de linkar categorias
     for(const category of categories.categories){
@@ -88,9 +74,9 @@ class MarketController {
    * @param {Request} ctx.request
    * @param {Response} ctx.response
    */
-  async show ({ params, request, response }) {
+  async show ({ params, response }) {
     const market_id = params.id
-    const market = await Market.find(market_id)
+    const market = await this.marketRepository.read(market_id)
 
     if( !market ) {
       return response.status(404).json({ message: 'Market was not found' })
@@ -105,14 +91,54 @@ class MarketController {
   }
 
   /**
+   * Update a single market
+   * PUT market/update/:id
+   *
+   * @param {object} ctx
+   * @param {Request} ctx.request
+   * @param {Response} ctx.response
+   */
+  async update({ params, request, response }) {
+    const market_id = params.id
+    const market = await this.marketRepository.update(market_id, request.body)
+
+    if( !market ){
+      return response.status(404).json({message:"Market wasn't found"})
+    }
+
+    if( request.body.address ){
+      const addressRepository = new AddressRepository()
+      await addressRepository.update(market.address_id, request.body.address)
+      market.address = await market.address().fetch()
+    }
+
+    if( request.body.categories ){
+      //Removendo os relacionamentos anteriores para inserir os novos
+      await market.categories().detach()
+
+      //Inserindo os novos relacionamentos
+      for(const category of request.body.categories ){
+        const categoryRegisted = await Category.query().where('category', category).first()
+
+        if( categoryRegisted ){
+          await market.categories().attach(categoryRegisted.id)
+        }
+      }
+
+      market.categories = await market.categories().fetch()
+    }
+
+    return { market }
+  }
+
+  /**
    * Delete a single market
    * POST market/delete/:id
    * Ok, it's working
    */
   async delete({ params, response }) {
-    const id = params.id
-
-    const market = await Market.find(id)
+    const market_id = params.id
+    const market = await this.marketRepository.read(market_id)
 
     if( !market ) {
       return response.status(404).json({ message: 'Market was not found' })
@@ -124,7 +150,7 @@ class MarketController {
     await address.delete()
     await market.delete()
 
-    return { message: "Tudo deletado" }
+    return response.status(200).json()
   }
 }
